+++ R/GetRegionOfInterest.R
@@ -4,8 +4,10 @@
 #'
 #' @param obj 'SpatialPoints*'.
 #'   Spatial points
-#' @param buffer 'numeric' number.
-#'   Buffer distance from convex hull.
+#' @param width 'numeric' number.
+#'   Buffer distance from geometry of convex hull.
+#' @param ...
+#'   Additional arguments to be passed to the \code{\link[rgeos]{gBuffer}} function.
 #'
 #' @return Returns an object of class 'SpatialPolygons'.
 #'
@@ -20,18 +22,18 @@
 #' @examples
 #' set.seed(123)
 #' pts <- sp::SpatialPoints(cbind(x = stats::runif(50), y = stats::runif(50)))
-#' sp::plot(GetRegionOfInterest(pts, buffer = 0.05), lty = 2)
+#' sp::plot(GetRegionOfInterest(pts, width = 0.05), lty = 2)
 #' sp::plot(GetRegionOfInterest(pts), add = TRUE)
 #' sp::plot(pts, col = "red", add = TRUE)
 #'
 
-GetRegionOfInterest <- function(obj, buffer=0) {
+GetRegionOfInterest <- function(obj, width=0, ...) {
   checkmate::assertClass(obj, "SpatialPoints")
-  checkmate::assertNumber(buffer, finite=TRUE)
+  checkmate::assertNumber(width, finite=TRUE)
 
   pts <- obj[grDevices::chull(sp::coordinates(obj)), ]
   ply <- sp::Polygons(list(sp::Polygon(pts)), ID=1)
   ply <- sp::SpatialPolygons(list(ply), proj4string=raster::crs(obj))
-  ply <- rgeos::gBuffer(ply, width=buffer)
+  ply <- rgeos::gBuffer(ply, width=width, ...)
   ply
 }